% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchGPT.R
\name{batch_mutate}
\alias{batch_mutate}
\title{Batch Mutate}
\usage{
batch_mutate(
  df,
  df_col,
  system_prompt,
  batch_size,
  batch_num,
  batch_total,
  model,
  temperature,
  rows
)
}
\arguments{
\item{df}{A data frame.}

\item{df_col}{A column from the data frame.}

\item{system_prompt}{A system prompt for the GPT model.}

\item{batch_size}{The current batch number tracked during processing.}

\item{model}{A GPT model. Default is "gpt-3.5-turbo-0125".}

\item{temperature}{A temperature for the GPT model. Default is 0.1.}

\item{rows}{A placeholder for the number of rows in the whole data frame.}
}
\value{
The final output data frame after processing all the rows.
}
\description{
This function applies the \code{mutate_row} function in batches to a data frame.
It processes a specified number of rows at a time and saves the progress.
It returns the final output data frame after all the rows have been processed.
}
