% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLLM_shiny.R
\name{batchLLM_shiny}
\alias{batchLLM_shiny}
\title{Interact with batchLLM via a Shiny Gadget}
\usage{
batchLLM_shiny()
}
\value{
No return value. Launches a Shiny Gadget that allows users to interact with the \code{batchLLM} package.
}
\description{
This function provides a user interface using Shiny to interact with
the \code{batchLLM} package. It allows users to configure and execute batch processing
through an interactive dashboard.
}
