% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beliefs.R
\docType{data}
\name{beliefs}
\alias{beliefs}
\title{Beliefs Dataset}
\format{
A data frame with 20 rows and 1 variable:
\describe{
\item{statement}{A character string with a statement representing a belief.}
}
}
\usage{
beliefs
}
\description{
The beliefs dataset consists of 20 statements representing opposing views on various scientific, environmental, and societal topics.
}
\examples{
head(beliefs)
}
\keyword{datasets}
