% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchGPT.R
\name{mutate_row}
\alias{mutate_row}
\title{Mutate Row}
\usage{
mutate_row(df, df_col, system_prompt, model, temperature)
}
\arguments{
\item{df}{A data frame.}

\item{df_col}{A column from the data frame.}

\item{system_prompt}{A system prompt for the GPT model.}

\item{model}{A GPT model. Default is "gpt-3.5-turbo-0125".}

\item{temperature}{A temperature for the GPT model. Default is 0.1.}
}
\value{
A data frame with the mutated row.
}
\description{
This function takes in a data frame \code{df}, a column from the data frame \code{df_col},
a system prompt \code{system_prompt}, a GPT model \code{model}, and a temperature \code{temperature}.
It uses the GPT model to generate a completion based on the system prompt and the content input.
The completion is assigned to the \code{gpt_output} column of the data frame.
}
