% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchGPT.R
\name{save_progress}
\alias{save_progress}
\title{Save Progress}
\usage{
save_progress(output, last_completed_batch)
}
\arguments{
\item{output}{The output to be saved.}

\item{last_completed_batch}{The last completed batch to be saved.}
}
\description{
This function takes in an output and a last completed batch,
and saves them as RDS files.
}
