% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLLM_shiny.R
\name{batchLLM_shiny}
\alias{batchLLM_shiny}
\title{Interact with batchLLM via a Shiny Gadget}
\usage{
batchLLM_shiny()
}
\description{
This function provides a Shiny-based user interface to interact with
the \code{batchLLM} package. It allows users to configure and execute batch processing
through an interactive dashboard.
}
