% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLLM_shiny.R
\name{batchLLM_shiny}
\alias{batchLLM_shiny}
\title{Launch a Shiny gadget that provides a user interface for batchLLM().}
\usage{
batchLLM_shiny()
}
\description{
Launch a Shiny gadget that provides a user interface for batchLLM().
}
