% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLLM.R
\name{claudeR}
\alias{claudeR}
\title{Interact with Anthropic's Claude API}
\usage{
claudeR(
  prompt,
  model = "claude-3-5-sonnet-20240620",
  max_tokens = 100,
  stop_sequences = NULL,
  temperature = 0.7,
  top_k = -1,
  top_p = -1,
  api_key = NULL,
  system_prompt = NULL
)
}
\arguments{
\item{prompt}{A string vector for Claude-2, or a list for Claude-3 specifying the input for the model.}

\item{model}{The model to use for the request. Default is the latest Claude-3 model.}

\item{max_tokens}{A maximum number of tokens to generate before stopping.}

\item{stop_sequences}{Optional. A list of strings upon which to stop generating.}

\item{temperature}{Optional.Amount of randomness injected into the response.}

\item{top_k}{Optional. Only sample from the top K options for each subsequent token.}

\item{top_p}{Optional. Does nucleus sampling.}

\item{api_key}{Your API key for authentication.}

\item{system_prompt}{Optional. An optional system role specification.}
}
\value{
The resulting completion up to and excluding the stop sequences.
}
\description{
This function was copied by yrvelez on GitHub (not currently available on CRAN).
}
